Solution5/backend/function_health/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
anyio==4.6.2.post1

azure-identity==1.17.1
azure-keyvault-secrets==4.9.0
azure-search-documents==11.6.0b6
azure-storage-blob==12.23.1
azure-data-tables==12.6.0
azure-servicebus==7.12.2

azure-storage-blob[aio]==12.23.1

httpx==0.27.2
pyodbc==5.1.0
